requirements.txt
customtkinter==5.0.3
matplotlib==3.6.2
numpy==1.21.5
opencv_python==4.7.0.68
pandas==1.5.2
Pillow==9.4.0
scipy==1.10.0
tkhtmlview==0.1.1.post5
